refactor more
