+++ app/src/pages/experiment/ExperimentRunLatency.tsx
@@ -1,0 +1,23 @@
+import { useMemo } from "react";
+
+import { LatencyText } from "@phoenix/components/trace/LatencyText";
+
+export function ExperimentRunLatency({
+  startTime,
+  endTime,
+}: {
+  startTime: string;
+  endTime: string;
+}) {
+  const latencyMs = useMemo(() => {
+    let latencyMs: number | null = null;
+    if (startTime && endTime) {
+      latencyMs = new Date(endTime).getTime() - new Date(startTime).getTime();
+    }
+    return latencyMs;
+  }, [startTime, endTime]);
+  if (latencyMs === null) {
+    return null;
+  }
+  return <LatencyText size="S" latencyMs={latencyMs} />;
+}
